Moved NamedConsequenceVisitor in a separate class
Move context outside of ModelGenerator

Extracted Visitor in NamedConsequenceVisitor.java

Moved getClassFromContext into Context
