+++ drools-model/drools-model-compiler/src/main/java/org/drools/modelcompiler/builder/generator/QueryParameter.java
@@ -1,0 +1,11 @@
+package org.drools.modelcompiler.builder.generator;
+
+public class QueryParameter {
+    final String name;
+    final Class<?> type;
+
+    public QueryParameter(String name, Class<?> type) {
+        this.name = name;
+        this.type = type;
+    }
+}
